math: lround assumed that long == int
There were *almost* enough casts in this function, except for one case
where converting a value that moved bits of the low-word of the
significand up to the high 32-bits of the dest.

+++ newlib/libm/common/s_lround.c
@@ -96,7 +96,7 @@ long int lround(double x)
 	/* 64bit long: shift amt in [32,42] */
         result = ((long int) msw << (exponent_less_1023 - 20))
 		/* 64bit long: shift amt in [0,10] */
-                | (lsw << (exponent_less_1023 - 52));
+            | ((long int) lsw << (exponent_less_1023 - 52));
       else
         {
 	  /* 32bit long: exponent_less_1023 in [20,30] */
